+++ app/src/main/java/com/greybox/projectmesh/db/MeshDatabase.kt
@@ -2,8 +2,8 @@ package com.greybox.projectmesh.db
 
 import androidx.room.Database
 import androidx.room.RoomDatabase
-import com.greybox.projectmesh.db.dao.MessageDao
-import com.greybox.projectmesh.db.entities.Message
+import com.greybox.projectmesh.messaging.data.dao.MessageDao
+import com.greybox.projectmesh.messaging.data.entities.Message
 
 @Database(entities = [Message::class], version = 2)
 abstract class MeshDatabase : RoomDatabase() {
